cts_recon_web/backend/requirements.txt
colorama
requests
beautifulsoup4
shodan
boto3
google-search-results
dnspython
flask
pymongo
python-dotenv
python-whois
wafw00f
